# packages/contracts/openapi.yaml
openapi: "3.0.3"
info:
  title: Engineering Modernization Hub API
  description: >
    modhub/v1 -- the single source of truth for the API surface consumed by
    the CLI, the Backstage backend plugin and lambda api itself. "El LLM
    propone, el nucleo dispone": this contract never lets a client set a
    run's final status directly -- only core_ops computes it.

    Operations carrying `x-modhub-status: designed` are part of the designed
    surface but are NOT deployed: they exist in neither lambda api nor API
    Gateway. Everything else is live. Keep this marker accurate -- it is the
    only thing separating the contract from a wish list.
  version: "1.1.0"
servers:
  - url: /modhub/v1
tags:
  - name: identity
  - name: strategies
  - name: runs

paths:
  /me:
    get:
      operationId: getMe
      x-modhub-status: designed
      tags: [identity]
      summary: The authenticated caller's identity.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Caller identity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /strategies:
    get:
      operationId: listStrategies
      x-modhub-status: designed
      tags: [strategies]
      summary: List registered strategy manifests.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Strategy catalog.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/StrategyManifest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /runs:
    post:
      operationId: createRun
      tags: [runs]
      summary: Create a modernization run.
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          x-modhub-status: designed
          schema:
            type: string
          description: >
            Designed, not implemented: the header is accepted and ignored,
            so two identical POSTs create two runs. Step Functions does
            reject duplicate execution names, but the name is the freshly
            generated run_id, so that protection sits in the wrong layer
            and never fires. Implementing it means a conditional write on
            (requested_by, key) before the run is created.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunRequest"
      responses:
        "201":
          description: Run created and its state-machine execution started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRunResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: >
            The free-text objetivo did not match any registered strategy
            within confidence. No run is created and no budget is spent.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    properties:
                      code:
                        type: string
                        enum: [NO_STRATEGY_MATCH]
    get:
      operationId: listRuns
      tags: [runs]
      summary: >
        List the caller's own runs. Always scoped to the `sub` in the token
        -- there is no parameter that widens it. A caller can only act on
        runs they requested, so listing anyone else's would return rows
        nothing can be done with.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          required: false
          description: Narrows the caller's own runs to one status.
          schema:
            $ref: "#/components/schemas/RunStatus"
        - in: query
          name: mine
          required: false
          deprecated: true
          description: >
            Accepted and ignored: the result is always the caller's own runs.
            Kept so existing clients do not break.
          schema:
            type: boolean
            default: true
        - in: query
          name: limit
          required: false
          x-modhub-status: designed
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: cursor
          required: false
          x-modhub-status: designed
          schema:
            type: string
      responses:
        "200":
          description: Page of runs.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Run"
                  next_cursor:
                    type: string
                    nullable: true
                    x-modhub-status: designed
        "401":
          $ref: "#/components/responses/Unauthorized"

  /runs/{run_id}:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      operationId: getRun
      tags: [runs]
      summary: >
        Read a run's current state. Scoped to the caller: a run requested by
        someone else answers 404, not 403 -- a 403 would confirm that the
        run_id exists.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /runs/{run_id}/events:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      operationId: getRunEvents
      x-modhub-status: designed
      tags: [runs]
      summary: Ordered event trail for a run.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: since_seq
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Events, ordered by sequence.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /runs/{run_id}/plan:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      operationId: getRunPlan
      x-modhub-status: designed
      tags: [runs]
      summary: The agent-proposed plan awaiting (or already given) approval.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The plan.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /runs/{run_id}/approval:
    parameters:
      - $ref: "#/components/parameters/RunId"
    post:
      operationId: approveRun
      tags: [runs]
      summary: >
        Approve or reject the pending plan. A single conditional write:
        requester's sub must match requested_by, state must be
        AWAITING_APPROVAL, plan_hash must match, auth must be recent.
        SendTaskSuccess/SendTaskFailure fires only if that write succeeds.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalRequest"
      responses:
        "200":
          description: Approval recorded; the run resumes or is rejected.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Caller is not the run's requester.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >
            plan_hash does not match the currently pending plan (it was
            superseded), or the run is not in AWAITING_APPROVAL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /runs/{run_id}/cancellation:
    parameters:
      - $ref: "#/components/parameters/RunId"
    post:
      operationId: cancelRun
      x-modhub-status: designed
      tags: [runs]
      summary: Cancel a run. A human decision, not a modernization outcome.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Run transitioned to CANCELADO.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Caller is not the run's requester.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Run is already in a terminal state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /runs/{run_id}/diff:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      operationId: getRunDiff
      x-modhub-status: designed
      tags: [runs]
      summary: Unified diff produced by the run, contained in approved paths.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The diff.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diff"
            text/plain:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /runs/{run_id}/pull-request:
    parameters:
      - $ref: "#/components/parameters/RunId"
    post:
      operationId: openPullRequest
      tags: [runs]
      summary: >
        Publish the run's result as a pull request on the source repository.
        The only component that writes outside the platform, which is why it
        is a separate Lambda rather than another branch inside lambda api.
        Explicit: nothing publishes automatically when a run finishes.
      description: >
        Idempotent. The resulting URL is stored on the run, so a second call
        returns the same pull request (200) instead of creating another. If
        the run has no stored URL but GitHub already has a pull request for
        the branch -- a crash between the two writes, or a run published
        before the URL was persisted -- that pull request is adopted and
        recorded rather than reported as an error.
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Pull request created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullRequest"
        "200":
          description: A pull request already existed for this run; the same one is returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Only the requester may publish this run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >
            The run is in a state with nothing to publish, or it changed no
            files.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: GitHub rejected the request; the message carries its response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /runs/{run_id}/report:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      operationId: getRunReport
      tags: [runs]
      summary: >
        Final report: verdict computed by core_ops from real JUnit, plus
        the agent's narrative (sources, decisions, rationale). The state
        never depends on the model's text. Scoped to the caller: a run
        requested by someone else answers 404, not 403.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The report.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    RunId:
      in: path
      name: run_id
      required: true
      schema:
        type: string
        format: uuid

  responses:
    Unauthorized:
      description: Missing or invalid JWT.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The run does not exist or is not visible to the caller.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    Error:
      type: object
      required: [code, message, run_id]
      properties:
        code:
          type: string
        message:
          type: string
        run_id:
          type: string
          format: uuid
          nullable: true

    Me:
      type: object
      required: [sub, email]
      properties:
        sub:
          type: string
        email:
          type: string
          format: email
        groups:
          type: array
          items:
            type: string

    RunStatus:
      type: string
      description: >
        PENDING, AWAITING_APPROVAL and RUNNING are in-flight. CANCELADO is a
        human decision, not a modernization outcome. The other five are the
        final states core_ops evaluates in strict order (first match wins):
        PRESUPUESTO_AGOTADO > BLOQUEADO > FALLIDO_CONTROLADO >
        COMPLETADO_PARCIALMENTE > LISTO_PARA_REVISION.
      enum:
        - PENDING
        - AWAITING_APPROVAL
        - RUNNING
        - CANCELADO
        - PRESUPUESTO_AGOTADO
        - BLOQUEADO
        - FALLIDO_CONTROLADO
        - COMPLETADO_PARCIALMENTE
        - LISTO_PARA_REVISION

    ResolvedStrategy:
      type: object
      required: [id, version, match_confidence]
      properties:
        id:
          type: string
        version:
          type: string
        match_confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1

    Restricciones:
      type: object
      properties:
        excluded_paths:
          type: array
          items:
            type: string
          default: []

    CreateRunRequest:
      type: object
      required: [repo, commit, objetivo, max_usd, max_iterations, max_minutes]
      properties:
        repo:
          type: string
          description: Repository URL.
        commit:
          type: string
          description: Exact commit SHA to fetch and run against.
          pattern: "^[0-9a-fA-F]{40}$"
        objetivo:
          type: string
          description: Free-text modernization objective.
          minLength: 1
        inputs:
          type: object
          description: Strategy-specific inputs (validated against the resolved strategy's manifest().inputs).
          additionalProperties: true
        restricciones:
          $ref: "#/components/schemas/Restricciones"
        max_usd:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
        max_iterations:
          type: integer
          minimum: 1
        max_minutes:
          type: integer
          minimum: 1

    CreateRunResponse:
      type: object
      required: [run_id, status, resolved_strategy, requested_by, created_at]
      properties:
        run_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/RunStatus"
        resolved_strategy:
          $ref: "#/components/schemas/ResolvedStrategy"
        requested_by:
          type: string
        created_at:
          type: string
          format: date-time

    Run:
      type: object
      required:
        - run_id
        - status
        - repo
        - commit
        - objetivo
        - requested_by
        - strategy_id
        - strategy_version
        - max_usd
        - max_iterations
        - max_minutes
        - spent_usd
        - iterations_used
        - created_at
        - updated_at
      properties:
        run_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/RunStatus"
        repo:
          type: string
        commit:
          type: string
          pattern: "^[0-9a-fA-F]{40}$"
        objetivo:
          type: string
        inputs:
          type: object
          additionalProperties: true
        restricciones:
          $ref: "#/components/schemas/Restricciones"
        requested_by:
          type: string
        strategy_id:
          type: string
        strategy_version:
          type: string
        max_usd:
          type: number
          format: float
        max_iterations:
          type: integer
        max_minutes:
          type: integer
        spent_usd:
          type: number
          format: float
        iterations_used:
          type: integer
        plan_hash:
          type: string
          nullable: true
        plan:
          type: object
          nullable: true
          additionalProperties: true
        changed_paths:
          type: array
          items:
            type: string
          description: Measured by core_ops from S3, never taken from the agent's account.
        pull_request_url:
          type: string
          format: uri
          nullable: true
          description: >
            Set once the run has been published. Its presence is what makes
            publishing idempotent -- a branch existing on GitHub is not a
            record the platform can read back.
        pull_request_branch:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Event:
      type: object
      required: [run_id, seq, type, created_at]
      properties:
        run_id:
          type: string
          format: uuid
        seq:
          type: integer
          minimum: 0
        type:
          type: string
          description: >
            e.g. RunCreated, PhaseStarted, PhaseCompleted, ApprovalRequested,
            ApprovalGranted, ApprovalRejected, SecurityBlocked, BudgetUpdated,
            RunCancelled, RunCompleted.
        message:
          type: string
        data:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time

    Plan:
      type: object
      required: [run_id, plan_hash, summary, steps, created_at]
      properties:
        run_id:
          type: string
          format: uuid
        plan_hash:
          type: string
        summary:
          type: string
        steps:
          type: array
          items:
            type: object
            required: [title]
            properties:
              title:
                type: string
              detail:
                type: string
        sources:
          type: array
          items:
            type: string
            format: uri
        created_at:
          type: string
          format: date-time

    ApprovalRequest:
      type: object
      required: [decision, plan_hash]
      properties:
        decision:
          type: string
          enum: [approve, reject]
        plan_hash:
          type: string
        reason:
          type: string

    Diff:
      type: object
      required: [run_id, unified_diff, writable_paths]
      properties:
        run_id:
          type: string
          format: uuid
        unified_diff:
          type: string
        writable_paths:
          type: array
          items:
            type: string

    PullRequest:
      type: object
      required: [run_id, pull_request_url, branch]
      properties:
        run_id:
          type: string
          format: uuid
        pull_request_url:
          type: string
          format: uri
        branch:
          type: string
        already_existed:
          type: boolean
          description: True when this call returned a pull request that already existed.

    Report:
      type: object
      description: >
        Two authors, one document (invariant #10). `verdict` is computed by
        core_ops from the real JUnit without Bedrock; `narrative` is the
        agent's prose. The status never depends on the model's text.
      required: [run_id, status, repo, commit, strategy, verdict, narrative, created_at]
      properties:
        run_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/RunStatus"
        repo:
          type: string
        commit:
          type: string
          pattern: "^[0-9a-fA-F]{40}$"
        objetivo:
          type: string
        strategy:
          type: object
          required: [id, version]
          properties:
            id:
              type: string
            version:
              type: string
        verdict:
          type: object
          required: [status, spent_usd, max_usd, iterations_used, max_iterations]
          properties:
            status:
              $ref: "#/components/schemas/RunStatus"
            spent_usd:
              type: number
              format: float
            max_usd:
              type: number
              format: float
            iterations_used:
              type: integer
            max_iterations:
              type: integer
        narrative:
          type: object
          description: Written by the agent. Never determines status.
          properties:
            summary:
              type: string
              nullable: true
            viability_reason:
              type: string
              nullable: true
            sources:
              type: array
              items:
                type: string
                format: uri
            risks:
              type: array
              items:
                type: string
        changed_paths:
          type: array
          items:
            type: string
        pull_request_url:
          type: string
          format: uri
          nullable: true
        diff:
          type: string
          nullable: true
          description: >
            Unified diff v0 -> v1, computed from S3. Truncated past 200000
            characters, with a trailing `[truncated]` marker -- a truncated
            diff will not apply cleanly.
        models_used:
          type: object
          additionalProperties:
            type: string
          description: Phase -> Bedrock model id. Platform configuration, never the agent's choice.
        plan_hash:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

    StrategyLimit:
      type: object
      required: [default, max]
      properties:
        default:
          type: number
        max:
          type: number

    StrategyLimits:
      type: object
      required: [max_usd, max_iterations, max_minutes]
      properties:
        max_usd:
          $ref: "#/components/schemas/StrategyLimit"
        max_iterations:
          $ref: "#/components/schemas/StrategyLimit"
        max_minutes:
          $ref: "#/components/schemas/StrategyLimit"

    StrategyManifest:
      type: object
      required: [id, version, title, checks, writable_paths, sources, limits, inputs]
      properties:
        id:
          type: string
        version:
          type: string
        title:
          type: string
        applies_to:
          type: object
          additionalProperties: true
        inputs:
          type: object
          description: JSON-schema-like description of the strategy's own inputs.
          additionalProperties: true
        limits:
          $ref: "#/components/schemas/StrategyLimits"
        checks:
          type: array
          items:
            type: string
        writable_paths:
          type: array
          items:
            type: string
        excluded_paths:
          type: array
          items:
            type: string
        sources:
          type: array
          items:
            type: string
            format: uri
        model_limits:
          $ref: "#/components/schemas/StrategyModelLimits"

    StrategyModelLimits:
      type: object
      description: Per-role max_tokens, fixed by the strategy (CLAUDE.md -- never chosen by the agent).
      properties:
        analysis_max_tokens:
          type: integer
          default: 4096
        code_max_tokens:
          type: integer
          default: 8192